requirements.txt
# Runtime dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
python-dateutil>=2.8.0
